backend/requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
msgpack==1.1.0
packaging==24.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
websockets==14.1
